add basic test code
