+++ app/src/main/java/com/nbusy/sdk/titan/neptulon/Message.java
@@ -2,48 +2,6 @@ package com.nbusy.sdk.titan.neptulon;
 
 import com.google.gson.JsonElement;
 
-/**
- * JSON-RPC request object.
- */
-class Request<T> {
-    final String id;
-    final String method;
-    final T params;
-
-    Request(String id, String method, T params) {
-        this.id = id;
-        this.method = method;
-        this.params = params;
-    }
-}
-
-/**
- * JSON-RPC response object.
- */
-class Response<T> {
-    final String id;
-    final T result;
-    final ResError error;
-
-    Response(String id, T result, ResError error) {
-        this.id = id;
-        this.result = result;
-        this.error = error;
-    }
-
-    class ResError<K> {
-        final int code;
-        final String message;
-        final K data;
-
-        ResError(int code, String message, K data) {
-            this.code = code;
-            this.message = message;
-            this.data = data;
-        }
-    }
-}
-
 
 /**
  * Generic (request or response) JSON-RPC message representation for incoming messages.
