+++ app/src/main/java/com/nbusy/sdk/titan/neptulon/Request.java
@@ -1,0 +1,16 @@
+package com.nbusy.sdk.titan.neptulon;
+
+/**
+ * JSON-RPC request object.
+ */
+public class Request<T> {
+    final String id;
+    final String method;
+    final T params;
+
+    Request(String id, String method, T params) {
+        this.id = id;
+        this.method = method;
+        this.params = params;
+    }
+}
